### 2021-05-21 #3304 @whedon
[PRE REVIEW]: LCIA Formatter
**Submitting author:** @WesIngwersen (<a href="http://orcid.org/0000-0002-9614-701X">Wesley Ingwersen</a>)
**Repository:** <a href="https://github.com/USEPA/LCIAformatter" target ="_blank">https://github.com/USEPA/LCIAformatter</a>
**Version:** v1.0.0
