### 2021-05-21 #3304 @kyleniemeyer
Hi @WesIngwersen, thanks for your interest in JOSS. Due to low editor availability in this area, I am placing your submission on a waitlist for now; once an appropriate editor becomes available, we will get this assigned and move the process forward.

In the meantime, you could address the issues with the references raised above, and also provide reviewer suggestions. Thanks!
