### 2021-05-21 #3304 @kyleniemeyer
@WesIngwersen regarding the paper generation error, I think you cannot have an empty `orcid` field for an author, so you should just remove that for someone that does not have an ORCID.

Also, I am not sure that the corresponding author footnote will work. @openjournals/dev is that correct?
