### 2021-05-21 4a43a0c
Remove empty orcid field for Michael
Per JOSS' whedon checks its not compiling. @msrocka If you have an ORCID and want to add it back, please feel welcome to do that
