### 2021-05-21 5ec5dde
Remove corresponding author tag
Per JOSS this might not compile with this note in the header YAML
