### 2021-05-21 #3304 @WesIngwersen
@kyleniemeyer, I removed the empty `orcid` field for Michael Srocka with [this commit](https://github.com/USEPA/LCIAformatter/commit/4a43a0c6ee7db3f254d69a109188ed030cdf264f)
and removed the corresponding author note with [this commit](https://github.com/USEPA/LCIAformatter/commit/5ec5dde087c17dea0f17b8dad9bfd16255c07ce9)
