### 2021-05-21 #3304 @danielskatz
You can use the command `@whedon check references` to check the references, and the command `@whedon generate pdf` to make a new PDF. Whedon commands need to be the first entry in a new comment.
