### 2021-05-21 #3304 @WesIngwersen
Recommended reviewer:

1. Chris Mutel, author of [Brightway: An open source framework for Life Cycle Assessment](https://joss.theoj.org/papers/10.21105/joss.00236) and two other JOSS articles
2. James Joyce, author of [Lcopt - An interactive tool for creating fully parameterised Life Cycle Assessment (LCA) foreground models](https://joss.theoj.org/papers/10.21105/joss.00339)
3. Konstantin Stadler, author of [The country converter coco - a Python package for converting country names between different classification schemes](https://joss.theoj.org/papers/10.21105/joss.00332)

All 3 recommended reviewers are PhD level domain experts (in life cycle assessment), Python developers, and have published in JOSS
