### 2021-05-21 #3304 @WesIngwersen
> Note that there are some potential reference problems flagged above too...

@danielskatz @kyleniemeyer  Regarding DOI issues.  I will address the missing DOI (actually others can be added).  The invalid DOIs are for 4 datasets that are data products of the LCIA formatter software and currently in the cue in the EPA data review/publication system. They should be through that process and registered within a week. I will report back on them once they are registered.
