### 2021-05-21 88a7172
-Add missing DOIs for journal articles
Add DOIs for Bare, Huijbreghts et al., and Bulle et al. refs.
Simplify URL for LCA ecosytem presentation
