+++ paper.bib
@@ -7,7 +7,8 @@
 	number = {5},
 	journaltitle = {Clean Technologies and Environmental Policy},
 	author = {Bare, Jane},
-	date = {2011}
+	date = {2011},
+	DOI = {10.1007/s10098-010-0338-9}
 }
 
 @article{huijbregts_recipe_2017,
@@ -17,7 +18,8 @@
 	number = {2},
 	journaltitle = {International Journal of Life Cycle Assessment},
 	author = {Huijbregts, M. A. J. and Steinmann, Z. J. N. and Elshout, P. M. F. and Stam, G. and Verones, F. and Vieira, M. D. M. and Hollander, A. and Zijp, M. and van Zelm, R.},
-	date = {2017}
+	date = {2017},
+	DOI = {10.1007/s11367-016-1246-y}
 }
 
 @article{bulle_impact_2019,
@@ -28,7 +30,8 @@
 	number = {9},
 	journaltitle = {The International Journal of Life Cycle Assessment},
 	author = {Bulle, Cécile and Margni, Manuele and Patouillard, Laure and Boulay, Anne-Marie and Bourgault, Guillaume and De Bruille, Vincent and Cao, Viêt and Hauschild, Michael and Henderson, Andrew and Humbert, Sebastien},
-	date = {2019}
+	date = {2019},
+	DOI = {10.1007/s11367-019-01583-0}
 }
 
  @book{goedkoop_recipe_2009, 
@@ -103,7 +106,7 @@
 
 @misc{ingwersen_lca_2019,
 	title = {{LCA} tool ecosystem},
-	url = {https://cfpub.epa.gov/si/si_public_record_report.cfm?dirEntryId=347759&Lab=CESER&simplesearch=0&showcriteria=2&sortby=pubDate&searchall=LCA+tool+ecosystem&timstype=&datebeginpublishedpresented=03/19/2016},
+	url = {https://cfpub.epa.gov/si/si_public_record_report.cfm?dirEntryId=347759},
 	urldate = {2021-03-19},
 	author = {Ingwersen, Wesley},
 	month = nov,
