### 2021-06-14 #3304 @WesIngwersen
@danielskatz  All DOIs are now active so reference issues are resolved.  FYI co-author @bl-young
