### 2021-06-20 #3304 @csoneson
👋🏻 @cmutel, @pjamesjoyce, @konstantinstadler - would (two of) you be willing to review this submission for JOSS?

> LCIA Formatter
> This paper describes a USEPA LCA ecosystem tool 'LCIA formatter' that extracts LCIA information from original source methods and converts the data for interoperability with the rest of the USEPA LCA ecosystem tools.
