### 2021-06-21 #3304 @csoneson
@cmutel - no problem, thanks for the quick response!

@mfastudillo @tngTUDOR @aleksandra-kim - would one or two of you be up for reviewing this submission for JOSS?

> LCIA Formatter
> This paper describes a USEPA LCA ecosystem tool 'LCIA formatter' that extracts LCIA information from original source methods and converts the data for interoperability with the rest of the USEPA LCA ecosystem tools.
