### 2021-06-22 #3392 @whedon
[REVIEW]: LCIA Formatter
**Submitting author:** @WesIngwersen (<a href="http://orcid.org/0000-0002-9614-701X">Wesley Ingwersen</a>)
**Repository:** <a href="https://github.com/USEPA/LCIAformatter" target ="_blank">https://github.com/USEPA/LCIAformatter</a>
**Version:** v1.0.2
