### 2021-06-25 USEPA/LCIAformatter#56 @mfastudillo
problems installing
Hi!

I am trying to install LCIAformatter with `pip install git+https://github.com/USEPA/LCIAformatter.git@v1.0.0#egg=LCIAformatter` in a new environment that has a freshly installed `fedelemflowlist`

I am getting these warnings and errors:

Collecting LCIAformatter
  Cloning https://github.com/USEPA/LCIAformatter.git (to revision v1.0.0) to c:\users\xxx\appdata\local\temp\pip-install-tk8wkmf2\lciaformatter_d086ec1651d9489fb51fcd412b56699c
  Running command git clone -q https://github.com/USEPA/LCIAformatter.git 'C:\Users\xxx\AppData\Local\Temp\pip-install-tk8wkmf2\lciaformatter_d086ec1651d9489fb51fcd412b56699c'
  Running command git checkout -q 4bf095d887de8dc3aed66656c12945b9d90952d8
  WARNING: Generating metadata for package LCIAformatter produced metadata for project name lciafmt. Fix your #egg=LCIAformatter fragments.
WARNING: Discarding git+https://github.com/USEPA/LCIAformatter.git@v1.0.0#egg=LCIAformatter. Requested lciafmt from git+https://github.com/USEPA/LCIAformatter.git@v1.0.0#egg=LCIAformatter has inconsistent name: filename has 'lciaformatter', but metadata has 'lciafmt'

ERROR: Could not find a version that satisfies the requirement lciaformatter (unavailable)
ERROR: No matching distribution found for lciaformatter (unavailable)

openjournals/joss-reviews#3392
