### 2021-06-25 USEPA/LCIAformatter#56 @bl-young
Sorry that may be the same issue: set the egg=lciafmt

pip install git+https://github.com/USEPA/LCIAformatter.git@v1.0.0#egg=lciafmt

I will update the wiki
