### 2021-06-26 USEPA/LCIAformatter#57 @mfastudillo
xlrd error getting methods
Hi!

I was trying the library and running `recipe2016 = lciafmt.get_method(lciafmt.Method.RECIPE_2016)` I got the following error

![imagen](https://user-images.githubusercontent.com/11621547/123506999-6cc8b300-d667-11eb-94e3-c283d6ae8ad2.png)

it seems to be caused by xlrd [ending support](https://groups.google.com/g/python-excel/c/IRa8IWq_4zk/m/Af8-hrRnAgAJ) for non-xls files. I installed pandas before lciafmt, and I ended up with version 2.0.1 fo xlrd... an alternative may be to switch to openpyx for xlsx files...

openjournals/joss-reviews#3392
