### 2021-06-26 USEPA/LCIAformatter#57 @bl-young
@mfastudillo yes it appears that the latest release of xlrd will not work. I will likely switch to openpyxl. In the meantime, you may want to downgrade to 1.2 to continue testing
