### 2021-06-29 b72b2c7
subset method for endpoints prior to mapping to avoid unclear warning; fillna so it doesn't get dropped during group_by
