### 2021-06-29 81638fe
enable naming of custom methods without needing a method_id
