+++ lciafmt/util.py
@@ -193,7 +193,13 @@ def generate_method_description(name: str) -> str:
         generic=yaml.safe_load(f)
     method_description = generic['description']
     method = check_as_class(name)
-    method_meta = method.get_metadata()
+    if method is None:
+        method_meta = {}
+        method_meta['name'] = name
+        method_meta['url'] = ''
+        method_meta['citation'] = ''
+    else:
+        method_meta = method.get_metadata()
     if 'detail_note' in method_meta:
         method_description += method_meta['detail_note']
     if 'methods' in method_meta:
@@ -262,11 +268,14 @@ def read_method(method_id):
     return method
 
 
-def save_json(method_id, mapped_data, method=None):
+def save_json(method_id, mapped_data, method=None, name = ''):
     """Saves a method as json file in the outputpath
     param method: str name of method to subset the passed mapped_data"""
     meta = set_lcia_method_meta(method_id)
-    filename = meta.name_data
+    if name == '':
+        filename = meta.name_data
+    else:
+        filename = name
     if method is not None:
         filename = method.replace('/','_')
         mapped_data = mapped_data[mapped_data['Method'] == method]
