### 2021-07-01 USEPA/LCIAformatter#56 @tngTUDOR
A quick "git clone" + "pip install ." worked fine for me.
