### 2021-07-02 6410043
specify xlrd requirement < 2.0
