### 2021-07-02 d496c0e
Update readme to include installation instructions
