### 2021-07-02 #3392 @bl-young
@mfastudillo @tngTUDOR
I'm making updates/edits in response to your issues raised on the [joss_review ](https://github.com/USEPA/LCIAformatter/tree/joss_review)branch

cc @wesingwersen
