### 2021-07-02 USEPA/LCIAformatter#57 @bl-young
requirements specified <2.0 for temporary resolution https://github.com/USEPA/LCIAformatter/commit/6410043d25b87cf6256743c25d90a5b27dafa397
