### 2021-07-03 USEPA/LCIAformatter#59 @mfastudillo
minor installation issue on linux machine
`pip install git+https://github.com/USEPA/LCIAformatter.git@joss_review#egg=lciafmt` throws an error when trying to install on a ubuntu machine. The problem seems to be that `pyodbc` would not install with  `pip install pyodbc` ... see [[pyodbc install instructions](https://github.com/mkleehammer/pyodbc/wiki/Install)]. I therefore installed first `unixodbc-dev` and then  `pip install git+https://github.com/USEPA/LCIAformatter.git@joss_review#egg=lciafmt` run apparently successfully. However if I try to import lciafmt from a ipython session I get the following error:

fatal: not a git repository (or any of the parent directories): .git

I am not sure how this problem can be handled, but I guess it can be at least noted on the installation instructions.

openjournals/joss-reviews#3392
