### 2021-07-03 USEPA/LCIAformatter#57 @mfastudillo
Hi

I ended up cloning the repo and using the setup file of the joss_review branch to install the library. When I try to get the TRACI method I have a different error.

![image](https://user-images.githubusercontent.com/11621547/124354711-9bb5ca80-dc0d-11eb-8937-468b24dea4c9.png)

the xlrd version installed is 1.2.0
