### 2021-07-05 USEPA/LCIAformatter#59 @tngTUDOR
My flow to install with Linux is as follows:

```
conda create -y -n lciafmt python=3.9
conda activate lciafmt
conda install pyodbc
pip install git+https://github.com/USEPA/LCIAformatter.git@joss_review#egg=lciafmt
```

And this install will work fine.
However, pyodbc installed on linux like this will not work. See below for the error message.

```
>>> lciafmt.clear_cache()
INFO delete cache folder /tmp/lciafmt
>>> impact_m = lciafmt.get_method(lciafmt.Method.ImpactWorld)
INFO get method ImpactWorld+
WARNING Please install drivers to remotely connect to Access Database. Drivers only available on windows platform.For instructions visit: https://github.com/mkleehammer/pyodbc/wiki/Connecting-to-Microsoft-Access
INFO download from https://www.dropbox.com/sh/2sdgbqf08yn91bc/AABIGLlb_OwfNy6oMMDZNrm0a/IWplus_public_v1.3.accdb?dl=1 to /tmp/lciafmt/Impact_World.accdb
INFO read ImpactWorld+ from file /tmp/lciafmt/Impact_World.accdb
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/maynardjk/workspaces/joss/LCIAformatter/lciafmt/__init__.py", line 81, in get_method
    return impactworld.get(file=file, url=url)
  File "/home/maynardjk/workspaces/joss/LCIAformatter/lciafmt/iw.py", line 36, in get
    df = _read(f)
  File "/home/maynardjk/workspaces/joss/LCIAformatter/lciafmt/iw.py", line 61, in _read
    cnxn = pyodbc.connect(connStr)

pyodbc.Error: ('01000', "[01000] [unixODBC][Driver Manager]Can't open lib 'Microsoft Access Driver (*.mdb, *.accdb)' : file not found (0) (SQLDriverConnect)")

```
