### 2021-07-14 USEPA/LCIAformatter#59 @bl-young
@mfastudillo

> fatal: not a git repository (or any of the parent directories): .git

This error occurs when you are not connected to a git - from this [block](https://github.com/USEPA/LCIAformatter/blob/master/lciafmt/util.py#L40-L44). Despite it being described as a fatal error, it has limited impact except that the git-hash won't be [appended to the file name](https://github.com/USEPA/esupy/wiki/File-Naming-Structure) of any method you produce. You should be able to proceed with all functions. It would be good to find a way to avoid that error, or at least explain it as such in the readme.
