### 2021-07-15 USEPA/LCIAformatter#59 @bl-young
@tngTUDOR
> WARNING Please install drivers to remotely connect to Access Database. Drivers only available on windows platform. For instructions visit: https://github.com/mkleehammer/pyodbc/wiki/Connecting-to-Microsoft-Access

We have added this warning to point users to this explanation for how to download the correct driver for pyodbc if you don't have it installed. However it seems this may not be feasible in linux. For ImpactWorld+, it might be limited to Windows only.
