### 2021-07-19 USEPA/LCIAformatter#59 @WesIngwersen
What about making pyodbc an optional dependency in the setup.py, like
[reference in setup tools docs](https://setuptools.readthedocs.io/en/latest/userguide/dependency_management.html)

```
setup(
...
extras_require={
        "ImpactWorld+":  ["pyodbc>=4.0.30;platform_system='Windows'"],
    }
}
```
As I interpret the docs, this setup will not install pyodbc by default with pip install.  You would have to use
>pip install LCIAformatter["ImpactWorld+"]

to install it, and still then it would only install pyodbc if it is a Windows machine due to the statement of that dependency afterwards.
