### 2021-07-19 USEPA/LCIAformatter#59 @WesIngwersen
Following that to also make pyodbc behave like an optional requirements in respect to `requirements.txt` we could move that line into an `ImpactWorld_requirements.txt` as pip permits install with multiple requirements files
