### 2021-07-19 USEPA/LCIAformatter#57 @WesIngwersen
It appears that we should try the alternative library proposed _openpyxl_ because of these versioning issues with _xlrd_
@bl-young
