### 2021-07-19 USEPA/LCIAformatter#57 @bl-young
I will work on resolving this issue to shift to openpyxl and will let you know when it has been fixed @mfastudillo @tngTUDOR. It seems that starting with python 3.9 xlrd will not work even with version 1.2
