+++ setup.py
@@ -12,9 +12,9 @@ setup(
                       "olca-ipc>=0.0.8",
                       "pandas>=0.22",
                       "xlrd>=1.2.0,<2.0.0",
-                      "pyyaml>=5.3",
-                      "pyodbc>=4.0.30",
+                      "pyyaml>=5.3"
                       ],
+    extras_require={"ImpactWorld": ["pyodbc >= 4.0.30;platform_system=='Windows'"]},
     license="MIT",
     author='Ben Young, Michael Srocka, Wesley Ingwersen, Troy Hottle, Ben Morelli',
     author_email='ingwersen.wesley@epa.gov',
