+++ impactworld_requirements.txt
@@ -1,0 +1,1 @@
+pyodbc>=4.0.30  #optional requirement for generating ImpactWorld method. Only works on Windows
