+++ requirements.txt
@@ -4,6 +4,5 @@ olca-ipc>=0.0.8
 pandas>=0.23
 requests>=2.21.0
 xlrd==1.2.0     # Not compatible with version 2.0.0
-pyodbc>=4.0.30
 pyyaml>=5.3
 
