### 2021-07-19 9c22542
Add more install instructions for optional requirements and note that ImpactWorld+ only works on Windows
- Addresses issue #59 along with previous two commits
