### 2021-07-19 USEPA/LCIAformatter#59 @WesIngwersen
@mfastudillo @tngTUDOR  I have implemented the changes I suggested in my comments. Can one of you test that it works on Linux and will not install pyodbc?  I hope this solution is acceptable to you given the constraints of the Impact World+ data being published as an Access db for this version. We can communicate with that team to note this limitation.
