### 2021-07-19 USEPA/LCIAformatter#59 @tngTUDOR
I agree  @WesIngwersen, it would be better for Open Source Initiatives if ImpactW+ was published
- in a more widespread db format (along MS access ) or at least something supporting Mac OS, Linux and Windows without too much trouble. Sqlite is the only thing coming to mind right now, given the context, but this depends on the features borrowed from Access. Of course, I would prefer feather or plain csv as last resource.
- definitely not in dropbox !

You have all my support to ask the IW+ team about this two things. Maybe they can use lciaformatter with a windows machine to directly publish a json-ld version on github
