### 2021-07-25 USEPA/LCIAformatter#59 @mfastudillo
Hello

I tried doing this on a linux machine :

`python3 -m venv eparev5`
`source eparev5/bin/activate`
`pip install git+https://github.com/USEPA/LCIAformatter.git@v1.0.0#egg=lciafmt`

and it ends up not using the wheel because it needs to be installed first. it does install pyodbc. If I install the wheels package before LCIAformatter, it also installs pyodbc. I'll try to restart everything from scratch

![image](https://user-images.githubusercontent.com/11621547/126895459-523780a0-f640-4696-a832-816e8af83082.png)
