### 2021-07-25 USEPA/LCIAformatter#59 @mfastudillo
I tried

```
python3 -m venv eparev6
source eparev6/bin/activate
pip install git+https://github.com/USEPA/LCIAformatter.git@joss_review#egg=lciafmt
```
and it does install lciafmt without installing pyodbc (although it complains of wheel not being installed). But if I try to import `lciafmt` I run into the same problem.

![image](https://user-images.githubusercontent.com/11621547/126905569-0b1f8ae1-6736-4b97-b45d-9389fbf62af3.png)

Nonetheless I can import the methods e.g. `lciafmt import get_method` and they seem to work fine
