### 2021-07-26 USEPA/LCIAformatter#59 @WesIngwersen
@mfastudillo  That error message is from trying to run a git command, so it seems `import lciafmt` is indirectly calling git. I'm guessing that is [here when we try to get a git hash for the current module](https://github.com/USEPA/LCIAformatter/blob/9c225425f6f62a47cfd7cf9f6f05f987db6d6ed8/lciafmt/util.py#L41), but it's inside a try clause, so I'm not sure why its returning that error. This is probably a separate issue.
