### 2021-07-27 5cc0d38
Move import of impact world module into get_method so it will only try to import it when called
