### 2021-07-29 24310bd
move reading of git_hash to esupy, new function checks for git presence prior to obtaining git_hash to avoid fatal error
