### 2021-07-30 f0b3110
update esupy requirements related to accessing git_hash (#60)
