### 2021-07-30 USEPA/LCIAformatter#59 @bl-young
The git hash issue in the most recent comment has been resolved (#60).

I believe this issue can also be closed with the updated install procedure for linux.
