### 2021-07-30 ed3b9c9
update handling of column headers, iter_cols cannot be used for read only workbooks; update use of cell_str
