### 2021-07-31 48903c2
refactor recipe midpoints with openpyxl
