+++ lciafmt/recipe.py
@@ -7,7 +7,6 @@ ReCiPe model
 """
 
 import pandas as pd
-import xlrd
 import openpyxl
 
 import lciafmt.cache as cache
@@ -120,7 +119,7 @@ def _read(file: str) -> pd.DataFrame:
     log.info("read ReCiPe 2016 from file %s", file)
     wb = openpyxl.load_workbook(file, read_only = True)
     records = []
-    for name in wb:
+    for name in wb.sheetnames:
         if _eqstr(name, "Version") or _eqstr(
                 name, "Midpoint to endpoint factors"):
             continue
@@ -132,7 +131,8 @@ def _read(file: str) -> pd.DataFrame:
 def _read_endpoints(file: str) -> pd.DataFrame:
     log.info("reading endpoint factors from file %s", file)
     wb = openpyxl.load_workbook(file, read_only = True)
-    endpoint_cols = ['Method','EndpointMethod', 'EndpointIndicator', 'EndpointUnit','EndpointConversion']
+    endpoint_cols = ['Method','EndpointMethod', 'EndpointIndicator',
+                     'EndpointUnit','EndpointConversion']
     endpoint = pd.DataFrame(columns = endpoint_cols)
     endpoints = []
     perspectives = ["I", "H", "E"]
@@ -146,7 +146,7 @@ def _read_endpoints(file: str) -> pd.DataFrame:
             flow_col = 0
 
             endpoint_factor_count = 0
-            for row in range(start_row, sheet.nrows):
+            for row in range(start_row, sheet.max_row):
                 indicator = xls.cell_str(sheet, row, flow_col)
                 indicator_unit = xls.cell_str(sheet, row, flow_col+1)
                 for i in range(0, 3):
@@ -185,11 +185,11 @@ def _read_endpoints(file: str) -> pd.DataFrame:
 
 def _read_mid_points(sheet: openpyxl.worksheet.worksheet.Worksheet,
                      records: list):
-    log.debug("try to read midpoint factors from sheet %s", sheet.name)
+    log.debug("try to read midpoint factors from sheet %s", sheet.title)
 
     start_row, data_col, with_perspectives = _find_data_start(sheet)
     if start_row < 0:
-        log.debug("could not find a value column in sheet %s", sheet.name)
+        log.debug("could not find a value column in sheet %s", sheet.title)
         return
 
     flow_col = _find_flow_column(sheet)
@@ -202,44 +202,44 @@ def _read_mid_points(sheet: openpyxl.worksheet.worksheet.Worksheet,
 
     perspectives = ["I", "H", "E"]
     factor_count = 0
-    for row in range(start_row, sheet.nrows):
+    for row in sheet.iter_rows(min_row=start_row):
         if compartment_col > -1:
-            compartment = xls.cell_str(sheet, row, compartment_col)
+            compartment = xls.cell_str(row[compartment_col])
         if compartment in contexts:
             compartment = contexts[compartment]
         if unit_col > -1:
-            flow_unit = xls.cell_str(sheet, row, unit_col)
+            flow_unit = xls.cell_str(row[unit_col])
             if "/" in flow_unit:
                 flow_unit = flow_unit.split("/")[1].strip()
         cas = ""
         if cas_col > -1:
-            cas=format_cas(xls.cell_f64(sheet, row, cas_col))
+            cas=format_cas(xls.cell_f64(row[cas_col]))
 
         if with_perspectives:
             for i in range(0, 3):
-                val = xls.cell_f64(sheet, row, data_col + i)
+                val = xls.cell_f64(row[data_col + i])
                 if val == 0.0:
                     continue
                 dfutil.record(records,
                               method="ReCiPe 2016 - Midpoint/" + perspectives[i],
[…]
