### 2021-07-31 18b3a4e
remove unneeded functions
