### 2021-07-31 90faa43
update requirements to openpyxl, remove xlrd
