### 2021-08-03 9413bb9
set data_only = True to read results of excel formulas
