### 2021-08-03 0d8f02d
access cas as value before formatting
