### 2021-08-03 cc70f1e
update esupy requirement to 0.1.5 to eliminate selenium sub-dependency
