+++ setup.py
@@ -8,7 +8,7 @@ setup(
     package_data={'lciafmt': ["data/*.*"]},
     include_package_data=True,
     install_requires=["fedelemflowlist @ git+git://github.com/USEPA/Federal-LCA-Commons-Elementary-Flow-List@v1.0.7#egg=fedelemflowlist",
-                      "esupy @ git+git://github.com/USEPA/esupy@v0.1.4#egg=esupy",
+                      "esupy @ git+git://github.com/USEPA/esupy@v0.1.5#egg=esupy",
                       "olca-ipc>=0.0.8",
                       "pandas>=0.22",
                       "openpyxl>=3.0.7",
