### 2021-08-18 #3392 @mfastudillo
>
>
> @tngTUDOR @mfastudillo We have tried resolved as best as we can the installation and use issues that you raised. Let us know if they are working now.

I've tried in a windows machine and everything seems to work fine
