### 2021-08-18 USEPA/LCIAformatter#62 @mfastudillo
mapping systems explanation
Hi

it would be nice if the "system" argument of the `lciafmt.map_flows` method is explained in more detail. The [documentation ](https://github.com/USEPA/LCIAformatter/wiki/Using-lciafmt)does not say much, nor the docstring of the method.

sounds kind of trivial, but shoud the user use the "source system" matching the method ?

e.g.
```
recipe2016 = lciafmt.get_method(lciafmt.Method.RECIPE_2016)
lciafmt.map_flows(recipe2016,system='ReCiPe2016')
```

openjournals/joss-reviews#3392
