### 2021-08-18 USEPA/LCIAformatter#63 @mfastudillo
community guidelines
as suggested by the template of joss, it would be good to have some basic guidelines for users wishing to 1) seek support 2) report issues

openjournals/joss-reviews#3392
