### 2021-08-19 ce60926
add warnings and exceptions if neither system or mapping is passed in map_flows
