+++ lciafmt/__init__.py
@@ -93,7 +93,9 @@ def to_jsonld(df: pd.DataFrame, zip_file: str, write_flows=False):
 
 def map_flows(df: pd.DataFrame, system=None, mapping=None,
               preserve_unmapped=False, case_insensitive=False) -> pd.DataFrame:
-    """Maps the flows in the given data frame using the given target system. It
+    """Maps the flows in the given data frame using the named target mapping
+       'system' from fedelemflowlist. Alternatively a mapping file can be passed using
+       'mapping' in the form of a dataframe meeting the mapping specifications. It
        returns a new data frame with the mapped flows."""
     mapper = fmap.Mapper(df, system=system, mapping=mapping,
                          preserve_unmapped=preserve_unmapped,
