### 2021-08-19 USEPA/LCIAformatter#62 @bl-young
I updated documentation on the wiki and updated the docstring for map_flows with https://github.com/USEPA/LCIAformatter/commit/ce60926308c2c073de14a715b5c44290aef277d6

Yes in general the system reflects the mapping file used from fedelemflowlist and the name will match the method. In the examples the name of the mapping file system is stored in [methods.json](https://github.com/USEPA/LCIAformatter/blob/master/lciafmt/data/methods.json#L7)
see: https://github.com/USEPA/LCIAformatter/blob/master/examples/traci.py#L24-L25

Let me know if this update helps.
