### 2021-08-23 #3392 @mfastudillo
a minor comment on the manuscript, when it says:

> LCA software providers of both proprietary and open source software have not provided an open process or external peer review of their incorporation of data from LCIA providers.

As far as I know brightway2 is totally open. Most of the impact assesssment methods come from ecoinvent ([e.g.](https://github.com/brightway-lca/brightway2-io/blob/master/bw2io/importers/ecoinvent_lcia.py)) and if the original sources are used, the code to process the sources is also open (https://github.com/brightway-lca/bw_recipe_2016). As far as I am aware this is an exception. OpenLCA [provides ](https://github.com/GreenDelta/data)the LCIA methods, but how it arrived there is not clear. It is probably worth to rephrase the sentence to make it more accurate.
