### 2021-08-24 USEPA/LCIAformatter#64 @WesIngwersen
Revise statement of lack of software LCIA incorporation transparency in paper.md
to include an exception for brightway
see https://github.com/openjournals/joss-reviews/issues/3392#issuecomment-904174107
