### 2021-08-31 07fdf84
add link in readme to wiki page for support #63
