### 2021-08-31 3e6b784
add brigthway exception to review of opensource LCIA methods #64
