+++ paper.bib
@@ -208,3 +208,17 @@
 	year = {2021}
 }
 
+@article{Mutel2017,
+  doi = {10.21105/joss.00236},
+  url = {https://doi.org/10.21105/joss.00236},
+  year = {2017},
+  publisher = {The Open Journal},
+  volume = {2},
+  number = {12},
+  pages = {236},
+  author = {Chris Mutel},
+  title = {Brightway: An open source framework for Life Cycle Assessment},
+  journal = {Journal of Open Source Software}
+}
+
+
