### 2021-09-22 cf35c88
remove mutable default arguments resolves #69
