### 2021-09-22 eefee66
revise olca-ipc requirements, 0.0.11 will require refactoring, resolves #65
