+++ requirements.txt
@@ -1,6 +1,6 @@
 git+git://github.com/USEPA/Federal-LCA-Commons-Elementary-Flow-List.git@v1.0.7#egg=fedelemflowlist
 git+git://github.com/USEPA/esupy.git@v0.1.5#egg=esupy
-olca-ipc>=0.0.8
+olca-ipc>=0.0.8, <=0.0.10
 pandas>=0.23
 requests>=2.21.0
 openpyxl>=3.0.7
