+++ lciafmt/__init__.py
@@ -27,50 +27,55 @@ class Method(Enum):
     RECIPE_2016 = "ReCiPe 2016"
     FEDEFL_INV = "FEDEFL Inventory"
     ImpactWorld = "ImpactWorld"
-    
+
     def get_metadata(cls):
         metadata = supported_methods()
         for m in metadata:
             if 'case_insensitivity' in m:
-                if m['case_insensitivity']=='True':
+                if m['case_insensitivity'] == 'True':
                     m['case_insensitivity'] = True
                 else:
                     m['case_insensitivity'] = False
             if m['id'] == cls.name:
                 return m
-    
+
     def get_filename(cls):
         filename = cls.get_metadata()['name'].replace(" ", "_")
         return filename
-    
+
     def get_path(cls):
         path = cls.get_metadata()['path']
         return path
 
     def get_class(name):
-        for n,c in Method.__members__.items():
+        for n, c in Method.__members__.items():
             m = c.get_metadata()
             mapping = None
             methods = {}
-            if 'mapping' in m: mapping = m['mapping']
-            if 'methods' in m: methods = m['methods']
+            if 'mapping' in m:
+                mapping = m['mapping']
+            if 'methods' in m:
+                methods = m['methods']
             if n == name or c.value == name or mapping == name or name in methods.keys():
                 return c
         util.log.error('Method not found')
 
+
 def supported_methods() -> list:
-    """Returns a list of dictionaries that contain meta-data of the supported
-       LCIA methods."""
+    """Return a dictionary of supported method meta_data."""
     json_file = pkg_resources.resource_filename("lciafmt", 'data/methods.json')
     with open(json_file, "r", encoding="utf-8") as f:
         return json.load(f)
 
 
-def get_method(method_id, add_factors_for_missing_contexts=True, endpoint=True, summary=False,
-               file=None, subset=None, url=None) -> pd.DataFrame:
-    """Returns the data frame of the method with the given ID. You can get the
-       IDs of the supported methods from the `supported_methods` function or
-       directly use the constants defined in the Method enumeration type."""
+def get_method(method_id, add_factors_for_missing_contexts=True,
+               endpoint=True, summary=False, file=None, subset=None,
+               url=None) -> pd.DataFrame:
+    """Generate the data frame of the method with the given ID.
+
+    The IDs of supported methods can be obtained using `supported_methods` or
+    directly use the constants defined in the Method enumeration type.
+    """
     method_id = util.check_as_class(method_id)
     if method_id == Method.TRACI:
         return traci.get(add_factors_for_missing_contexts, file=file, url=None)
@@ -82,21 +87,27 @@ def get_method(method_id, add_factors_for_missing_contexts=True, endpoint=True, 
     if method_id == Method.FEDEFL_INV:
         return fedefl_inventory.get(subset)
 
+
 def clear_cache():
     cache.clear()
 
+
 def to_jsonld(df: pd.DataFrame, zip_file: str, write_flows=False):
-    """Generates a JSONLD file of the methods passed as dataframe."""
+    """Generate a JSONLD file of the methods passed as dataframe."""
     util.log.info("write JSON-LD package to %s", zip_file)
     with jsonld.Writer(zip_file) as w:
         w.write(df, write_flows)
 
+
 def map_flows(df: pd.DataFrame, system=None, mapping=None,
               preserve_unmapped=False, case_insensitive=False) -> pd.DataFrame:
-    """Maps the flows in the given data frame using the named target mapping
-       'system' from fedelemflowlist. Alternatively a mapping file can be passed using
-       'mapping' in the form of a dataframe meeting the mapping specifications. It
-       returns a new data frame with the mapped flows."""
+    """Map the flows in a method using a mapping from fedelemflowlist.
+
[…]
