### 2021-09-22 1727956
expand function doc strings for init
