### 2021-09-22 470980f
add more function annotations
