### 2021-09-23 73e46ee
update doc strings and linting, update to f strings
