+++ lciafmt/util.py
@@ -29,7 +29,7 @@ datapath = modulepath + '/data/'
 log.basicConfig(level=log.INFO, format='%(asctime)s %(levelname)-8s %(message)s',
                 datefmt='%Y-%m-%d %H:%M:%S', stream=sys.stdout)
 
-#Common declaration of write format for package data products
+# Common declaration of write format for package data products
 write_format = "parquet"
 
 paths = Paths()
@@ -38,6 +38,15 @@ outputpath = paths.local_path
 
 pkg = pkg_resources.get_distribution('lciafmt')
 git_hash = get_git_hash()
+
+method_metadata = {
+    'Name': '',
+    'Version': '',
+    'Source': '',
+    'SourceType': '',
+    'Citation': '',
+    }
+
 
 def set_lcia_method_meta(method_id):
     lcia_method_meta = FileMeta()
@@ -53,13 +62,6 @@ def set_lcia_method_meta(method_id):
     lcia_method_meta.git_hash = git_hash
     return lcia_method_meta
 
-method_metadata = {
-    'Name':'',
-    'Version':'',
-    'Source':'',
-    'SourceType':'',
-    'Citation':'',
-    }
 
 def make_uuid(*args: str) -> str:
     path = _as_path(*args)
@@ -82,20 +84,11 @@ def is_non_empty_str(s: str) -> bool:
     return s.strip() != ""
 
 
-def is_empty_str(s: str) -> bool:
-    if s is None:
-        return True
-    if isinstance(s, str):
-        return s.strip() == ''
-    else:
-        return False
-
-
 def format_cas(cas) -> str:
-    """ In LCIA method sheets CAS numbers are often saved as numbers. This
-        function formats such numbers to strings that matches the general
-        format of a CAS numner. It also handles other cases like None values
-        etc."""
+    """In LCIA method sheets CAS numbers are often saved as numbers. This
+    function formats such numbers to strings that matches the general
+    format of a CAS numner. It also handles other cases like None values.
+    """
     if cas is None:
         return ""
     if cas == "x" or cas == "-":
@@ -117,9 +110,9 @@ def aggregate_factors_for_primary_contexts(df) -> pd.DataFrame:
     :param df: a pandas dataframe for an LCIA method
     :return: a pandas dataframe for an LCIA method
     """
-    #Ignore the following impact categories for generating averages
+    # Ignore the following impact categories for generating averages
     ignored_categories = ['Land transformation', 'Land occupation',
-                          'Water consumption','Mineral resource scarcity',
+                          'Water consumption', 'Mineral resource scarcity',
                           'Fossil resource scarcity']
     indices = df['Context'].str.find('/')
     ignored_list = df['Indicator'].isin(ignored_categories)
@@ -140,13 +133,14 @@ def aggregate_factors_for_primary_contexts(df) -> pd.DataFrame:
         primary_context.append(sub)
 
     df['Primary Context'] = primary_context
-    #Subset the df to only include the rows were a primary context was added
+    # Subset the df to only include the rows were a primary context was added
     df_secondary_context_only = df[df['Primary Context'].notnull()]
 
-    #Determine fields to aggregate over. Do not use flow UUID or old context
-    agg_fields = list(set(df.columns) - {'Context', 'Flow UUID', 'Characterization Factor'})
+    # Determine fields to aggregate over. Do not use flow UUID or old context
+    agg_fields = list(set(df.columns) - {'Context', 'Flow UUID',
+                                         'Characterization Factor'})
 
-    #drop primary context field from df
+    # drop primary context field from df
     df = df.drop(columns=['Primary Context'])
 
     df_secondary_agg = df_secondary_context_only.groupby(agg_fields, as_index=False).agg(
[…]
