+++ lciafmt/endpoint.py
@@ -11,23 +11,25 @@ import pandas as pd
 import lciafmt
 from .util import log
 
+
 def apply_endpoints(endpoints, matching_fields):
-    """
-    Returns a dataframe in LCIAmethod format that contains endpoint factors
-        based on conversion factors supplied in passed param 'endpoints'
-    param endpoints must conform to Endpoints specs
-    param matching_fields: list of fields on which to apply unique endpoint
+    """Generate an endpoint method in LCIAmethod format.
+
+    Method contains endpoint factors based on conversion factors supplied in
+    passed dataframe.
+    :param endpoints: df which conforms to Endpoints specs
+    :param matching_fields: list of fields on which to apply unique endpoint
         conversions
     """
     log.info('developing endpoint methods...')
 
     indicators = endpoints[['Method'] + matching_fields]
-    
+
     method = pd.DataFrame()
-    
+
     for e in matching_fields:
         endpoints[e].fillna("", inplace=True)
-    
+
     for m in indicators['Method'].unique():
         method_indicators = indicators[indicators['Method']==m]
         mapped_method = lciafmt.get_mapped_method(m, methods=[m])
@@ -54,11 +56,11 @@ def apply_endpoints(endpoints, matching_fields):
         agg_fields = list(set(endpoint_method.columns) - {'Characterization Factor'})
         endpoint_method_agg = endpoint_method.groupby(agg_fields, as_index=False).agg(
             {'Characterization Factor': 'sum'})
-        
+
         # Sort
         endpoint_method_agg = endpoint_method_agg[endpoint_method.columns]
         endpoint_method_agg.sort_values(by=['Indicator','Flowable','Context'], inplace=True)
-        
+
         method = method.append(endpoint_method_agg, ignore_index = True)
-    
+
     return method

[…]
