+++ lciafmt/endpoint.py
@@ -31,22 +31,21 @@ def apply_endpoints(endpoints, matching_fields):
         endpoints[e].fillna("", inplace=True)
 
     for m in indicators['Method'].unique():
-        method_indicators = indicators[indicators['Method']==m]
+        method_indicators = indicators[indicators['Method'] == m]
         mapped_method = lciafmt.get_mapped_method(m, methods=[m]).fillna('')
         if 'Indicator' in matching_fields:
             mapped_method = mapped_method[mapped_method['Indicator'].isin(
                 list(method_indicators['Indicator']))]
-        endpoint_method = mapped_method.merge(endpoints[matching_fields +
-                                                  ['Method',
-                                                  'Endpoint Indicator',
-                                                  'Endpoint Indicator unit',
-                                                  'Conversion factor']],
-                                       how = 'left',
-                                       on = matching_fields + ['Method'])
-        endpoint_method['Indicator']=endpoint_method['Endpoint Indicator']
-        endpoint_method['Indicator unit']=endpoint_method['Endpoint Indicator unit']
-        endpoint_method['Characterization Factor']=endpoint_method['Characterization Factor']\
-            *endpoint_method['Conversion factor']
+        endpoint_method = mapped_method.merge(
+            endpoints[matching_fields + ['Method', 'Endpoint Indicator',
+                                         'Endpoint Indicator unit',
+                                         'Conversion factor']],
+            how='left',
+            on=matching_fields + ['Method'])
+        endpoint_method['Indicator'] = endpoint_method['Endpoint Indicator']
+        endpoint_method['Indicator unit'] = endpoint_method['Endpoint Indicator unit']
+        endpoint_method['Characterization Factor'] = endpoint_method['Characterization Factor']\
+            * endpoint_method['Conversion factor']
         endpoint_method.drop(columns=['Endpoint Indicator',
                                       'Endpoint Indicator unit',
                                       'Conversion factor'],
@@ -55,15 +54,16 @@ def apply_endpoints(endpoints, matching_fields):
         if (len(endpoint_method.index) != len(mapped_method.index)):
             log.warn("some characterization factors lost")
 
-        #Determine fields to aggregate over. Use all fields except CF which are summed
+        # Determine fields to aggregate over. Use all fields except CF which are summed
         agg_fields = list(set(endpoint_method.columns) - {'Characterization Factor'})
         endpoint_method_agg = endpoint_method.groupby(agg_fields, as_index=False).agg(
             {'Characterization Factor': 'sum'})
 
         # Sort
         endpoint_method_agg = endpoint_method_agg[endpoint_method.columns]
-        endpoint_method_agg.sort_values(by=['Indicator','Flowable','Context'], inplace=True)
+        endpoint_method_agg.sort_values(by=['Indicator', 'Flowable',
+                                            'Context'], inplace=True)
 
-        method = method.append(endpoint_method_agg, ignore_index = True)
+        method = method.append(endpoint_method_agg, ignore_index=True)
 
     return method
