+++ setup.py
@@ -2,13 +2,13 @@ from setuptools import setup
 
 setup(
     name="lciafmt",
-    version="1.0.0",
+    version="1.0.1",
     packages=["lciafmt"],
     package_dir={'lciafmt': 'lciafmt'},
     package_data={'lciafmt': ["data/*.*"]},
     include_package_data=True,
-    install_requires=["fedelemflowlist @ git+git://github.com/USEPA/Federal-LCA-Commons-Elementary-Flow-List@v1.0.7#egg=fedelemflowlist",
-                      "esupy @ git+git://github.com/USEPA/esupy@v0.1.5#egg=esupy",
+    install_requires=["fedelemflowlist @ git+git://github.com/USEPA/Federal-LCA-Commons-Elementary-Flow-List#egg=fedelemflowlist",
+                      "esupy @ git+git://github.com/USEPA/esupy#egg=esupy",
                       "olca-ipc>=0.0.8, <=0.0.10",
                       "pandas>=0.22",
                       "openpyxl>=3.0.7",
