### 2021-10-06 4684b40
👷 Implement basic testing with gh actions
