+++ .github/workflows/python-package.yml
@@ -5,9 +5,8 @@ name: Python package
 
 on:
   push:
-    branches: [ master, joss_review, develop ]
   pull_request:
-    branches: [ master ]
+    branches: [master, develop]
 
 jobs:
   build:
