### 2021-10-06 635263a
👷 Rename action
previous name referred to packaging, but it's actually about testing.
