+++ .github/workflows/python-package.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Python package
+name: Python testing
 
 on:
   push:
@@ -9,7 +9,7 @@ on:
     branches: [master, develop]
 
 jobs:
-  build:
+  test_linux:
 
     runs-on: ubuntu-latest
     strategy:
