### 2021-10-06 d65088d
💚 Always install requirements.txt
The repo package use a requirements to declare dependencies, no need to test for it's existance.
