+++ .github/workflows/python-package.yml
@@ -67,6 +67,6 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        pip install .
+        pip install .[ImpactWorld]
         pytest
 
