+++ README.md
@@ -12,14 +12,15 @@ source, cleans the data, shapes them into a standard format using the [LCIAmetho
 |ImpactWorld+ Midpoint*|International Reference Center for Life Cycle of Products, Services and Systems (CIRAIG)|[ImpactWorld+](http://www.impactworldplus.org/en/team.php)|
 |ImpactWorld+ Endpoint*|International Reference Center for Life Cycle of Products, Services and Systems (CIRAIG)|[ImpactWorld+](http://www.impactworldplus.org/en/team.php)|
 |FEDEFL Inventory Methods|US Environmental Protection Agency|[FEDEFL Inventory Methods](https://github.com/USEPA/LCIAformatter/wiki/Inventory-Methods)|
-* only works on Windows installations
+
+\* only works on Windows installations
 
 ## Installation Instructions
 
 Install a release directly from github using pip. From a command line interface, run:
-> pip install git+https://github.com/USEPA/LCIAformatter.git@v1.0.0#egg=lciafmt
+> pip install git+https://github.com/USEPA/LCIAformatter.git@v1.0.1#egg=lciafmt
 
-where you can replace 'v1.0.0' with the version you wish to use under [Releases](https://github.com/USEPA/LCIAformatter/releases).
+where you can replace 'v1.0.1' with the version you wish to use under [Releases](https://github.com/USEPA/LCIAformatter/releases).
 
 Alternatively, to install from the most current point on the repository:
 ```
