### 2021-10-07 97452b8
add additional comments to install testing yml
