### 2021-10-08 97278b7
Minor fixes to references+typos
