+++ paper.md
@@ -39,15 +39,15 @@ Life Cycle Assessment (LCA) is an established and standardized methodology to co
 # Statement of need
 A simplified algorithm for LCA is given in [@eq:lca], where $I$ are impacts, $E$ are emissions generated (e.g. pollutants) or raw resources consumed (e.g. land, water) per functional unit of product across the entire supporting product system and $CF$ are corresponding characterization factors that quantitatively relate a unit of a flow to a given impact indicator (e.g., global warming, acidification, total land use, etc.). 
 
-$$
+\begin{equation}\label{eq:lca}
 I = \sum(E*CF)
-$$ {#eq:lca}
+\end{equation}
 
 Both $E$ and $CF$ use objects called elementary flows, which are data objects that generally represent a substance (e.g. Ammonia), source or receiving environmental context (e.g. Freshwater lake), and unit (e.g. kilogram). $E$ will have a total quantity of a given elementary flow from the product system, and the characterization factors in $CF$ are in the form of the indicator unit per elementary flow unit (e.g. kg N-eq per kg Ammonia). $E$ comes from the LCI calculated for the given product under user-defined conditions, and $CF$ is a static dataset that comes from an LCIA method provider. 
 
 LCA software generally include LCIA methods to provide impact assessment results for user-created and software-provided LCI, but the elementary flows used in these LCIA methods must be the same data objects as the flows in the LCI data to enable proper impact result calculations. A major challenge to reproducing LCA results and sharing LCA models across platforms is that there is no internationally common list of elementary flows used by either LCI, LCIA, or software providers. A critical review of elementary flow data used in LCA showed that flows in LCIA methods are the least clearly described and therefore least portable and machine-readable among those from LCI, LCIA and software providers [@edelen_critical_2018]. Furthermore, LCA software providers of both proprietary and open source software rarely provide an open process or external peer review of their incorporation of data from LCIA providers (Brightway2 is one exception [@Mutel2017]). This is one potential cause of discrepancies in results across LCA software.
 
-The LCIA formatter is a specific solution to this problem. The LCIA formatter transparently acquires LCIA methods from original provider data portals, maps them to an authoritative flow list, and exports them in common data formats. The LCIA formatter v1.0 uses the Federal LCA Commons Elementary Flow List (FEDEFL) as the authoritative system of elementary flows [@edelen_federal_2019]. This system has been adopted by federal agencies in the U.S. to share data through the Federal LCA Commons [(FLCAC)](https://www.lcacommons.gov/). The automation of this process provided by the LCIA formatter is critical, because as elementary flows used in LCI are updated, the LCIA datasets should also be automatically updated, and vice versa, to facilitate their use in LCA. 
+The LCIA formatter is a specific solution to this problem. The LCIA formatter transparently acquires LCIA methods from original provider data portals, maps them to an authoritative flow list, and exports them in common data formats. The LCIA formatter v1.0 uses the Federal LCA Commons Elementary Flow List (FEDEFL) as the authoritative system of elementary flows [@edelen_federal_2019]. This system has been adopted by federal agencies in the U.S. to share data through the Federal LCA Commons ([FLCAC](https://www.lcacommons.gov/)). The automation of this process provided by the LCIA formatter is critical, because as elementary flows used in LCI are updated, the LCIA datasets should also be automatically updated, and vice versa, to facilitate their use in LCA. 
 
 # Structure
[…]
