+++ paper.md
@@ -50,7 +50,7 @@ LCA software generally include LCIA methods to provide impact assessment results
 The LCIA formatter is a specific solution to this problem. The LCIA formatter transparently acquires LCIA methods from original provider data portals, maps them to an authoritative flow list, and exports them in common data formats. The LCIA formatter v1.0 uses the Federal LCA Commons Elementary Flow List (FEDEFL) as the authoritative system of elementary flows [@edelen_federal_2019]. This system has been adopted by federal agencies in the U.S. to share data through the Federal LCA Commons ([FLCAC](https://www.lcacommons.gov/)). The automation of this process provided by the LCIA formatter is critical, because as elementary flows used in LCI are updated, the LCIA datasets should also be automatically updated, and vice versa, to facilitate their use in LCA. 
 
 # Structure
-The LCIA formatter code is written in Python 3 and created as a standard python package called `lciafmt` that can be installed using pip. The LCIA formatter primarily uses pandas for data wrangling [@mckinney-proc-scipy-2010], the Apache parquet format [@apache_pyarrow_2021] for local storage of processed datasets, and olca-ipc [@srocka_olca-ipc_2020] for writing data in a standard LCA data format. The code is stored on a USEPA GitHub [repository](https://github.com/USEPA/LCIAformatter) and is available for public access.
+The LCIA formatter code is written in Python 3 and created as a standard python package called `lciafmt` that can be installed using pip. The LCIA formatter primarily uses pandas [@mckinney-proc-scipy-2010] for data wrangling, the Apache parquet format [@apache_pyarrow_2021] for local storage of processed datasets, and olca-ipc [@srocka_olca-ipc_2020] for writing data in a standard LCA data format. The code is stored on a USEPA GitHub [repository](https://github.com/USEPA/LCIAformatter) and is publically accessible.
 
 The LCIA formatter accesses LCIA datasets directly from the data provider. These datasets are typically provided as Microsoft Excel or Access files. These are downloaded and saved in a temporary local cache.
 To support the specific functions necessary to access and parse individual methods, each method is processed within its own Python module. Flow names, indicators, characterization factors, and other metadata are compiled in a [standard format](https://github.com/USEPA/LCIAformatter/tree/documentation/format%20specs).
@@ -59,7 +59,7 @@ Additionally, the LCIA formatter supports the inclusion of non-specified seconda
 Where methods provide both midpoint and endpoint categories within a single source, the LCIA formatter parses these indicators for separate use.  Midpoint indicators are typically calculated at an intermediate point along the cause-effect model (such as final mass of substance in the environment), and endpoint indicators typically correspond to damage from the elementary flows (such as a disability-adjusted life year, or DALY) [@unep_glam_2016].  Therefore, midpoint and endpoint indicators should be evaluated separately.
 Finally, source flow data are mapped to elementary flows in the FEDEFL [@edelen_federal_2019], through mapping files provided within that [package](https://github.com/USEPA/Federal-LCA-Commons-Elementary-Flow-List) [@ingwersen_fedelemflowlist_2020]. These mapping files correspond to flow names and contexts to a common set of elementary flows generated for LCA modeling by the USEPA. Because flows are provided at more detailed contexts than exists in most LCIA methods currently, mapping to the FEDEFL typically expands the number of characterization factors available.
 Mapped methods are stored locally as parquet files for future access by LCIA formatter or other tools.
-Additionally, mapped methods can be exported as JSON-LD format for use in LCA software tools such as [openLCA](https://www.openlca.org/).
+Additionally, mapped methods can be exported in JSON-LD format for use in LCA software tools such as [openLCA](https://www.openlca.org/).
 
 
 # Available Methods
