+++ .github/workflows/python-package.yml
@@ -7,8 +7,18 @@ name: Python install test
 on:
   push:
     branches: [master, develop]
+    paths-ignore:
+      - '**.md'
+      - '**.bib'
+      - 'LICENSE'
+      - '.gitignore'
   pull_request:
     branches: [master, develop]
+    paths-ignore:
+      - '**.md'
+      - '**.bib'
+      - 'LICENSE'
+      - '.gitignore'
   workflow_dispatch:  # also allow manual trigger, for testing purposes
 
 jobs:
