### 2021-10-08 #3392 @bl-young
Tagged release: https://github.com/USEPA/LCIAformatter/releases/tag/v1.0.2
